Back/trade

See merge request s09-final/S09P31C202!93

+++ workspace/backend/mpick/src/main/java/com/k9c202/mpick/trade/entity/Trade.java
@@ -45,10 +45,10 @@ public class Trade {
     private String tradeExplain;
 
     @Column(name = "count")
-    private Integer wishCount;
+    private Long wishCount;
 
     @Column
-    private Integer viewCount;
+    private Long viewCount;
 
     @Column
     @Enumerated(EnumType.ORDINAL)
